Refactor LogAndReturn method to CompatibilityTests for better encapsulation
- Move LogAndReturn from CustomComparator to CompatibilityTests class
- Remove verboseMode parameter, use class field VerboseMode directly
- Improve code organization by keeping test-specific logic in test class
- Maintain all existing property-based test functionality
- Reorganize CustomComparator into Helper directory for better structure
